requirements.txt
aniso8601==9.0.1
attrs==21.4.0
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
docopt==0.6.2
filelock==3.7.1
finnhub-python==2.4.13
Flask==2.1.2
Flask-RESTful==0.3.9
gevent==21.12.0
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.1
pandas==1.4.3
parsel==1.6.0
pipreqs==0.4.11
Protego==0.2.1
psycopg==3.0.15
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
queuelib==1.6.2
requests==2.28.0
requests-file==1.5.1
Scrapy==2.6.1
scrapyrt==0.13.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
tldextract==3.3.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.9
w3lib==1.22.0
websocket==0.2.1
websocket-client==1.3.3
Werkzeug==2.1.2
yarg==0.1.9
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
